+++ services/search/pkg/models.go
@@ -20,7 +20,7 @@ type LocationRadius struct {
 }
 
 type RestaurantSearchResult struct {
-	ID         int32    `json:"id"`
+	ID         int64    `json:"id"`
 	Name       string   `json:"name"`
 	Latitude   float64  `json:"latitude"`
 	Longitude  float64  `json:"longitude"`
